+++ src/App.css
@@ -1,13 +1,11 @@
-.App {
-  text-align: center;
-}
-
 .App-logo {
   height: 5vmin;
   pointer-events: none;
 }
 
-.App-body {
+.App {
+  text-align: center;
+  font-family: 'Overpass', sans-serif;
   background-color: #003046;
   min-height: 100vh;
   display: flex;
@@ -22,6 +20,24 @@
   color: #61dafb;
 }
 
+.image-contact {
+max-width: 60%;
+height: auto;
+align-items: center;
+}
+
+.image-logo {
+max-width: 50%;
+height: auto;
+align-items: center;
+}
+
+.footer{
+  font-family: 'Overpass', sans-serif;
+  font-size: calc(5px + 1vmin);
+  position: fixed;
+  bottom: 0;
+}
 
 li {
   text-align:left;
@@ -31,18 +47,25 @@ p{
   text-align:left;
 }
 
+h2{
+  text-align:center;
+}
 
 h3{
   text-align:left;
 }
 
+
 a:link {
 color: white;
 font-weight: bold;
+text-decoration: none;
 }
 a:visited {
 color: white;
+text-decoration: none;
 }
 a:hover {
-color: #4E5BCC;
+color: #white;
+text-decoration: none;
 }
