+++ src/components/Footer.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+function Footer(){
+  return (
+
+    <div className="footer">
+      <p><center>Copyright © 2022 A/B Street</center></p>
+    </div>
+
+  );
+}
+
+export default Footer;
